Problem DCP-196 is ready to update.

+++ DevSkillCodingProblem/DCP-196.cs
@@ -8,12 +8,13 @@ namespace DevSkillCodingProblem
     {
         static void Main(string[] args)
         {
-            var T = Convert.ToInt16(Console.ReadLine());
+            var T = Convert.ToInt32(Console.ReadLine());
             for (int i = 0; i < T; i++)
             {
                 var v = Convert.ToDecimal(Console.ReadLine());
                 var t = Convert.ToDecimal(Console.ReadLine());
                 int r = (int)(v / t);
+
                 Console.WriteLine(string.Format("-{0:n2}", r));
             }
         }
